Move query params test to separate module
